3.3-permissions/api_with_restrictions/requirements.txt
django==4.0.5
djangorestframework
psycopg2-binary
django-filter
